pyproject.toml
[tool.poetry]
name = "barcodeserver"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.112.0"}
python-barcode = {extras = ["images"], version = "^0.15.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
